requirements.txt
playwright >= 1.37
pytest-playwright
pytesseract >= 0.3.10
Pillow >= 10.0.0python-dotenv >= 1.0.0
